--- src/modules/dataBase.js
@@ -1,2 +1,0 @@
-let db;
-export default db;
