+++ src/modules/db.js
@@ -1,0 +1,4 @@
+const idList = ['52982', '52854', '52861', '52806', '52881', '52883', '52846', '52927', '52928', '52912'];
+const apiBaseUrl = 'https://www.themealdb.com/api/json/v1/1/lookup.php?i=';
+
+export { idList, apiBaseUrl };
